fix(ci): first-party line limit — split oversized ancestry callees tests over 1500 lines
